Create event.py
Not sure this is the right thing to do, but in this event.py file there is only a shared variable to store the event system for the entire app. I tried to follow the solution proposed with functions, and this is the result. In the only-functions solution the event system is in a .py file, in which it is directly stored a subscribers dictionary.